requirements.txt
Flask==2.0.2
gunicorn==20.1.0
flask-cors==3.0.10
python-dotenv==0.19.2
Werkzeug==2.0.3
openai
